Fix | Otel Adapter Config (#95)
* Fix | Update OTELAdapter to start collector if not running during config update

* Fix | Ensure OTEL collector service is reset after errors in Initialize and StartAgent methods

+++ agent/internal/adapters/otel_adapter.go
@@ -55,6 +55,9 @@ func (a *OTELAdapter) Initialize() error {
 		if err := a.svc.Run(a.ctx); err != nil {
 			logger.Logger.Sugar().Errorf("OTEL collector stopped with error: %v", err)
 		}
+		a.mu.Lock()
+		a.svc = nil
+		a.mu.Unlock()
 	}()
 	return nil
 }
@@ -79,6 +82,9 @@ func (a *OTELAdapter) StartAgent() error {
 		if err := a.svc.Run(a.ctx); err != nil {
 			logger.Logger.Sugar().Errorf("OTEL collector stopped with error: %v", err)
 		}
+		a.mu.Lock()
+		a.svc = nil
+		a.mu.Unlock()
 	}()
 	return nil
 }
@@ -99,27 +105,32 @@ func (a *OTELAdapter) StopAgent() error {
 
 func (a *OTELAdapter) UpdateConfig() error {
 	a.mu.RLock()
-	defer a.mu.RUnlock()
+	running := a.svc != nil
+	a.mu.RUnlock()
+
+	if !running {
+		if err := a.StartAgent(); err != nil {
+			return fmt.Errorf("failed to start OTEL collector: %w", err)
+		}
+		logger.Logger.Info("OTEL collector started with updated config")
+		return nil
+	}
 
 	sigChan := make(chan os.Signal, 1)
 	signal.Notify(sigChan, syscall.SIGHUP)
 	defer signal.Stop(sigChan)
-
-	go func() {
-		select {
-		case sig := <-sigChan:
-			logger.Logger.Sugar().Infof("Received signal for updating config in otel collector: %s\n", sig)
-		case <-time.After(5 * time.Second):
-			logger.Logger.Warn("Timeout waiting for config update signal")
-		}
-	}()
 
 	logger.Logger.Info("Sending SIGHUP signal to hot-reload otel collector for updating config...")
 	if err := syscall.Kill(os.Getpid(), syscall.SIGHUP); err != nil {
 		return fmt.Errorf("failed to send SIGHUP signal: %w", err)
 	}
 
-	time.Sleep(2 * time.Second)
+	select {
+	case sig := <-sigChan:
+		logger.Logger.Sugar().Infof("Received signal for updating config in otel collector: %s", sig)
+	case <-time.After(5 * time.Second):
+		logger.Logger.Warn("Timeout waiting for config update signal")
+	}
 
 	logger.Logger.Info("Config updated. OTEL collector restarted")
 	return nil
